+++ EliteAPI/Inara/Events/DelCommanderInventoryCargoItem.cs
@@ -9,7 +9,7 @@ namespace EliteAPI.Inara.Events
 {
     public class DelCommanderInventoryCargoItem : IInaraEventData
     {
-        public DelCommanderInventoryCargoItem(string itemName, int itemCount)
+        public DelCommanderInventoryCargoItem(string itemName, long itemCount)
         {
             ItemName = itemName;
             ItemCount = itemCount;
@@ -19,7 +19,7 @@ namespace EliteAPI.Inara.Events
         public string ItemName { get; set; }
 
         [JsonProperty("itemCount")]
-        public int ItemCount { get; set; }
+        public long ItemCount { get; set; }
 
         [JsonProperty("isStolen")]
         public bool? IsStolen { get; set; }
